use navbar component
